@@ -6,41 +6,77 @@
     "type": "git",
     "url": "https://github.com/Herrad/neoscan"
   },
-  "main": "index.js",
+  "main": "main.js",
   "scripts": {
-    "test": "mocha tests/**/*.js",
-    "build": "rollup -c"
+    "build": "rollup -c",
+    "electron": "electron electron/main.js",
+    "start": "electron-forge start",
+    "package": "electron-forge package",
+    "make": "electron-forge make",
+    "publish": "electron-forge publish"
   },
-  "bin": {
-    "neoscan": "bin/neoscan"
-  },
-  "author": "",
+  "author": "herrad.koe@gmail.com",
   "license": "ISC",
   "dependencies": {
-    "cli-table": "^0.3.1",
-    "colors": "^1.4.0",
-    "commander": "^5.0.0",
-    "cookie-parser": "^1.4.6",
-    "express": "^4.18.1",
-    "hbs": "^4.2.0",
-    "http-errors": "^2.0.0",
-    "morgan": "^1.10.0",
-    "open": "^8.4.0",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
-    "tail": "^2.0.3",
-    "ws": "^8.7.0"
+    "electron-squirrel-startup": "^1.0.0",
+    "electron-store": "^8.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "tail": "^2.2.4"
   },
   "devDependencies": {
     "@babel/core": "^7.18.2",
     "@babel/preset-env": "^7.18.2",
     "@babel/preset-react": "^7.17.12",
+    "@electron-forge/cli": "^6.0.0-beta.64",
+    "@electron-forge/maker-deb": "^6.0.0-beta.64",
+    "@electron-forge/maker-rpm": "^6.0.0-beta.64",
+    "@electron-forge/maker-squirrel": "^6.0.0-beta.64",
+    "@electron-forge/maker-zip": "^6.0.0-beta.64",
+    "@electron-forge/publisher-github": "^6.0.0-beta.64",
     "@rollup/plugin-babel": "^5.3.1",
     "@rollup/plugin-commonjs": "^22.0.0",
     "@rollup/plugin-node-resolve": "^13.3.0",
     "@rollup/plugin-replace": "^4.0.0",
-    "mocha": "^7.1.1",
-    "rollup": "^2.75.4",
-    "should": "^11.1.1"
+    "electron": "^19.0.6",
+    "rollup": "^2.75.4"
+  },
+  "config": {
+    "forge": {
+      "packagerConfig": {},
+      "publishers": [
+        {
+          "name": "@electron-forge/publisher-github",
+          "config": {
+            "repository": {
+              "owner": "herrad",
+              "name": "neoscan"
+            }
+          }
+        }
+      ],
+      "makers": [
+        {
+          "name": "@electron-forge/maker-squirrel",
+          "config": {
+            "name": "neoscan"
+          }
+        },
+        {
+          "name": "@electron-forge/maker-zip",
+          "platforms": [
+            "darwin"
+          ]
+        },
+        {
+          "name": "@electron-forge/maker-deb",
+          "config": {}
+        },
+        {
+          "name": "@electron-forge/maker-rpm",
+          "config": {}
+        }
+      ]
+    }
   }
 }
