@@ -1,9 +1,9 @@
 import React from 'react';
 
 const Resistance = ({ reduction, totalReducedDamage } = []) => (
-    <span className="resistance">
+    <li className="resistance">
         {reduction.source} {reduction.reducedDamageBy.toFixed(2)} ({(reduction.reducedDamageBy / totalReducedDamage * 100).toFixed(0)}%)
-    </span>
+    </li>
 )
 
 export default Resistance
