@@ -3,7 +3,5 @@ import ReactDOM from 'react-dom/client';
 import App from './app';
 
 const rootElement = document.getElementById('root')
-if (rootElement) {
-  const root = ReactDOM.createRoot(rootElement);
-  root.render(<App />);
-}
+const root = ReactDOM.createRoot(rootElement);
+root.render(<App />);
